Model/requirements.txt
stable-baselines3
gym
torch
numpy
pandas
ta
pyzmq
